Make use of environment variables to determine usage of RabbitMQ, Kafka or RestAPI in cert_signer and certificate_api
