+++ backend/certificate_api/configs/config.js
@@ -14,6 +14,11 @@ OQIDAQAB
 const KAFKA_BOOTSTRAP_SERVER = process.env.KAFKA_BOOTSTRAP_SERVERS || 'localhost:9092';
 const RABBITMQ_SERVER = process.env.RABBITMQ_SERVER || 'localhost:5672';
 const EVENTS_TOPIC = 'events';
+const COMMUNICATION_MODE_KAFKA = "kafka";
+const COMMUNICATION_MODE_RABBITMQ = "rabbitmq";
+const COMMUNICATION_MODE_RESTAPI = "restapi";
+const COMMUNICATION_MODE = COMMUNICATION_MODE_RABBITMQ;
+
 module.exports = {
     REGISTRY_URL,
     JWT_PUBLIC_KEY,
@@ -21,5 +26,9 @@ module.exports = {
     RABBITMQ_SERVER,
     EVENTS_TOPIC,
     KEYCLOAK_URL,
-    KEYCLOAK_REALM
+    KEYCLOAK_REALM,
+    COMMUNICATION_MODE_KAFKA,
+    COMMUNICATION_MODE_RABBITMQ,
+    COMMUNICATION_MODE_RESTAPI,
+    COMMUNICATION_MODE
 };
