+++ backend/certificate_signer/config/config.js
@@ -22,6 +22,11 @@ const CERTIFICATE_RETRY_COUNT = process.env.CERTIFICATE_RETRY_COUNT || 5;
 const REDIS_URL = process.env.REDIS_URL || 'redis://0.0.0.0:6379';
 const REDIS_KEY_EXPIRE = process.env.REDIS_KEY_EXPIRE || 2 * 24 * 60 * 60; // in secs
 
+const COMMUNICATION_MODE_KAFKA = "kafka";
+const COMMUNICATION_MODE_RABBITMQ = "rabbitmq";
+const COMMUNICATION_MODE_RESTAPI = "restapi";
+const COMMUNICATION_MODE = COMMUNICATION_MODE_RABBITMQ;
+
 module.exports = {
   CERTIFICATE_NAMESPACE,
   CERTIFICATE_CONTROLLER_ID,
@@ -45,5 +50,8 @@ module.exports = {
   REDIS_URL,
   REDIS_KEY_EXPIRE,
   DUPLICATE_CERTIFICATE_TOPIC,
-
+  COMMUNICATION_MODE_KAFKA,
+  COMMUNICATION_MODE_RABBITMQ,
+  COMMUNICATION_MODE_RESTAPI,
+  COMMUNICATION_MODE
 };
